add Cart API

+++ DataAccess/Migrations/DataContextModelSnapshot.cs
@@ -103,8 +103,8 @@ namespace DataAccess.Migrations
                     b.Property<int>("TotalItem")
                         .HasColumnType("int");
 
-                    b.Property<int>("TotalPrice")
-                        .HasColumnType("int");
+                    b.Property<double>("TotalPrice")
+                        .HasColumnType("float");
 
                     b.Property<int>("UserId")
                         .HasColumnType("int");
